Remove unnecessary layer
For now, we only have one option which provides the file path for the
tokenizer. And along with this line, we don't dispatch any other functions from
the output provided by cli_parser. Also, about custom error messages, when an option
marked required in parser, it automatically raises its own errors. That means it's
already handled.