Add mock packer dispatcher
The logic is, when there's a -p info, it'll send "True" param to
the obfuscator and, it'll immediately create the context of the
file.
The logic of the dispatcher is simple, it'll check the packer names
and dispatch whichever that selected by the user.

Though the implementation is required to be polish, like there
should be a control flow that checks if packer value is send or
something along with these lines.